rename: application 패키지명을 service로 변경
