+++ src/main/java/com/web/stard/domain/study/api/StudyController.java
@@ -1,6 +1,6 @@
 package com.web.stard.domain.study.api;
 
-import com.web.stard.domain.member.application.MemberService;
+import com.web.stard.domain.member.service.MemberService;
 import com.web.stard.domain.member.domain.Member;
 import com.web.stard.domain.study.domain.dto.StudyRequestDto;
 import com.web.stard.domain.study.domain.dto.StudyResponseDto;
@@ -9,7 +9,6 @@ import com.web.stard.domain.study.service.StudyService;
 import com.web.stard.domain.study.service.StudyTagService;
 import com.web.stard.global.domain.CurrentMember;
 import jakarta.validation.Valid;
-import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
